// desktop/src/open-api/model/recognitionTaskStatus.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type RecognitionTaskStatus = 'AWAITING_UPLOAD' | 'UPLOADING' | 'UPLOAD_INTERRUPTED' | 'DISPATCH_PENDING' | 'QUEUED' | 'RUNNING' | 'RETRY_WAIT' | 'SUCCEEDED' | 'FAILED';

export const RecognitionTaskStatus = {
    AwaitingUpload: 'AWAITING_UPLOAD' as RecognitionTaskStatus,
    Uploading: 'UPLOADING' as RecognitionTaskStatus,
    UploadInterrupted: 'UPLOAD_INTERRUPTED' as RecognitionTaskStatus,
    DispatchPending: 'DISPATCH_PENDING' as RecognitionTaskStatus,
    Queued: 'QUEUED' as RecognitionTaskStatus,
    Running: 'RUNNING' as RecognitionTaskStatus,
    RetryWait: 'RETRY_WAIT' as RecognitionTaskStatus,
    Succeeded: 'SUCCEEDED' as RecognitionTaskStatus,
    Failed: 'FAILED' as RecognitionTaskStatus
};